backend/pyproject.toml
[tool.ruff]
# Target Python version
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Also enable isort (`I`) and flake8-bugbear (`B`).
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008", "E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.migrations.*"
ignore_errors = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpath = ["."]
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
asyncio_mode = "auto"
env = [
    "DATABASE_URL=postgresql+asyncpg://user:password@localhost:5432/outputudy",
    "SECRET_KEY=test_secret_key",
    "GCP_PROJECT_ID=test-project",
    "ENV=test",
]

[project]
name = "outputudy-backend"
version = "0.1.0"
description = "Outputudy Backend"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.0.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.10",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]
